+++ SMS_Form/Model/Mark.cs
@@ -12,7 +12,7 @@ namespace SMS_Form.Model
         public int StudentId { get; set; }
         
         public int ExamId { get; set; }
-        public double Marks { get; set; }
+        public int Marks { get; set; }
         public string StudentName { get; set; }
        
         public string ExamName { get; set; }
